Add new unified buffer interface

+++ cfunc/include/util.h
@@ -45,7 +45,7 @@ template<typename T>
 void assignValIfEmpty(std:: vector<T> & v1, std::vector<T> const &v_assign, int start_dim, T default_val) {
     //function to assign the vector element from start_dim to end,
     //if it's empty assign default_val
-    assert(start_dim <= v_assign.size() &&
+    assert(start_dim <= (int)v_assign.size() &&
             "assign dimension should not exceed the target vector dimension!\n");
     v1.assign(v_assign.begin()+start_dim, v_assign.end());
     if (v1.empty())
